New installer type: easy_install
This is mainly used to install .exe installers for Python modules
generated by distutils.

Pip is not able to silently install this kind of packages but
easy_install is able to unwrap the ZIP archive contained within
these .exe files and install them silently.

To use it simply specify an installer of type 'easy_install_26'
for Python 2.6 and 'easy_install_27' for Python 2.7